// src/app/core/models/table/extras_column.enum.ts
export enum ExtrasColumn {
  ROUNDED_CIRCLE = 'rounded_circle',
  ICON = 'icon',
  BINARY = 'binary',
  REVERSE_BINARY = 'reverse_binary',
  CURRENCY = 'currency',
  DATE = 'date',
  PERCENTAGE = 'percentage',
  BADGE = 'badge',
  SWITCH = 'switch',
  IMAGE = 'image',
  LINK = 'link',
  VOUCHER = 'voucher',
  APPEND = 'append',
  DEPEND = 'depend',
  SIZE_FILE = 'sizeFile',
  PIPE = 'pipe',
  ACTION_ICON = 'actionIcon',
  VALUE_FIXED = 'valueFixed',
  LIST_VALUE = 'list',
}