// src/app/core/models/table/table-type.enum.ts
export enum TableType {
  INCOMING_CORRESPONDENCE = 'incoming_correspondence',
  ARCHIVED_CORRESPONDENCE = 'archived_correspondence',
  DISTRIBUTED_CORRESPONDENCE = 'distributed_correspondence',
  OUTGOING_CORRESPONDENCE = 'outgoing_correspondence',
  ACCEPTED_MAIL_LIST = 'accepted_mail_list',
  REGISTRY_REQUEST = 'registry_request',
  ORGANIZATION = 'organization',
  SUB_ORGANIZATION = 'sub-organization',
  COUNTRY = 'country',
  MEMBER = 'Member',
  GROUP = 'Group',
  USERS = 'users',
  RECYCLE_BIN = 'recycle_bin',
  MAIL_TEMPLATE = 'mail-template',
  CORRESPONDENCE_FOLLOW_UP = 'correspondence-follow-up',
}

export enum ServiceType {
  INCOMING_CORRESPONDENCE = 'bureau-v2/courrier/paginate/entrant',
  ARCHIVED_CORRESPONDENCE = 'bureau-v2/courrier/paginate/archiver',
  OUTGOING_CORRESPONDENCE = 'bureau-v2/courrier/paginate/sortant',
  DISTRIBUTED_CORRESPONDENCE = 'bureau-v2/courrier/paginate/distribuer',
  ACCEPTED_MAIL_LIST = 'bureau-v2/demande/demande-acceptee',
  REGISTRY_REQUEST = 'bureau-v2/demande/demande-bureauordre',
  REGISTRY_REQUEST_EMPLOYER = 'bureau-v2/demande/demande-bureauordre/simple-user',
  ORGANIZATION = 'ref-v2/organisation',
  SUB_ORGANIZATION = 'ref-v2/sous-organisation',
  COUNTRY = 'ref-v2/pays',
  MEMBER = 'ref-v2/membre',
  GROUP = 'ref-v2/groupe-membre',
  USERS = 'auth-v2/users',
  RECYCLE_BIN = 'bureau-v2/courrier/paginate/reject',
  MAIL_TEMPLATE = 'ref-v2/modele/paginate',
  CORRESPONDENCE_FOLLOW_UP = 'bureau-v2/courrier/paginate/suivi-courrier',
}